Place repaint boundaries around fancy buttons in color tests
We're still not passing on Travis. The golden file tests work great on my
machine, so I'm trying to reduce the problem space as much as I can think to do.
This will omit the scaffold from the goldens. We'll see what happens.

I'm anticipating another failure. I wonder if Travis can upload what it sees to
imgur or somesuch.