Break out icon and label widgets into a separate, stateless container
I have the idea to create a "core" widget that represents the state of the whole
icon-label assembly at a single instant in time. Animation will be kept entirely
in FancyButtonState, but transition tweens will be computed in the stateless
core widget. A single "progress" value will be passed to the constructor of the
core to facilitate this, as well as "from-" and "to-poses".